Automake is a jealous god.  If, when sacrificing a goat to it, you do not kill the goat with the *correct* sort of knife, on the *correct* altar, and drink its blood from the *correct* goblet at the *correct* temperature, he will wreak his revenge, perhaps, for example, causing all your crops to sprout bearing bright green and pink polka-dots.
Add an extra 3 microns to the tip of the blade, so that the Automake God
will not be angry and fail to generate rules to install man pages.  (It
would have been nice had the sacred scrolls described that particular
incantation, but so it goes....)

While we're at it, get rid of "EXTRA_MANS"; it doesn't appear to be
necessary (but such hubris may, of course, bring down the wrath of the
Automake God upon me - you can't just get rid of EXTRA_PROGRAMS, for
example, as the Automake God then forgets how to handle "ethereal_LDADD"
and the like - although in a test run the generated Makefile.in didn't
*appear* to have anything missing other than a definition of EXTRA_MANS,
which it didn't use and so presumably wouldn't miss).


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@2900 f5534014-38df-0310-8fa8-9805f1628bb7