Remove some more "CHECK_DISPLAY_AS_DATA()" calls and code to set "pinfo->current_proto", in dissectors always called through dissector tables and handles.
Make the IEEE 802.11 dissector static, as it's not called externally.

Clear the Info column in the Linux cooked capture and 802.1q VLAN
dissectors, before extracting anything from the packet, so that if an
exception is thrown, the Info column doesn't reflect the previous
protocol.

Don't extract the encapsulated protocol in the VLAN dissector until you
use it, so that if the frame contains the VLAN ID but not the
encapsulated protocol, we at least put the VLAN ID into the protocol
tree.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@2927 f5534014-38df-0310-8fa8-9805f1628bb7