Remove more "CHECK_DISPLAY_AS_DATA()" calls and "pinfo->current_proto =" statements.
Move the setting of the Protocol column in various dissectors before
anything is fetched from the packet, and also clear the Info column at
that point in those and some other dissectors, so that if an exception
is thrown, the columns don't reflect the previous protocol.

Make the IP dissector static, as it's called only via dissector tables
or dissector handles.  Also make the "dissect the TOS field as the
DiffServ DS field" flag static, as it's not referred to outside of
"packet-ip.c".

In the NCP dissector, refer to the port type through "pinfo" rather than
through the global "pi", as it's a tvbuffified dissector.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@2929 f5534014-38df-0310-8fa8-9805f1628bb7