Calculate the height and width of m_r_font globally, since various routines need it.
When a user clicks on a hex digit or on the corresponding character
(the "text dump" portion) in the hex dump, find the field in the
proto_tree that the byte corresponds to, expand the GtkCTree so that
the field is viewable, select the field, and center it vertically.
LanAlyzer has this feature, and I've missed it in Ethereal.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@3096 f5534014-38df-0310-8fa8-9805f1628bb7