When dissecting LANMAN pipe transaction replies, store with each reply frame per-frame data indicating
	1) what type of transaction it's a reply to

and

	2) whether it's the first reply or a continuation reply

as the information supplied by the SMB dissector can only be trusted on
the first pass through the capture.

(If you have two different transactions in the *same* conversation with
the *same* MID, but different transaction types, only on the first pass
will the transaction type in the data structure pointed to by
"si.request_val" reflect the previous request - it reflects the last
request seen which, when the user is clicking on frames in the capture,
needn't be the request corresponding to the reply that they've just
clicked on.

If you have a reply that consists of multiple SMBs,
"si.request_val->trans_response_seen" will be set to 1 as soon as the
first reply is seen, and will *remain* 1 until the request is seen
again; if the user clicks on one of the SMBs in the reply, even if it's
the first SMB in the reply, without having first clicked on the request,
"si.request_val->trans_response_seen" will be 1, and the SMB will be
misdissected as a continuation.)

Use common code to handle the beginning of LANMAN replies, rather than
duplicating it in the code to handle each reply.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@3154 f5534014-38df-0310-8fa8-9805f1628bb7