Make "comp_info_str" static in Ethereal - there's no need for it outside "gtk/main.c" in Ethereal.
Add the GLib version to it in Ethereal, and put in the GLib version
rather than the GTK+ version in Tethereal (which isn't linked with
GTK+...).

Make it a GString; this makes the code to construct it slightly less
ugly, especially now that we're putting the GLib version in.

Fix the code for the "-D" flag in Tethereal to compile in a no-libpcap
version (in a no-libpcap version, it just says that this version of
Tethereal wasn't compiled with capture support).


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@3196 f5534014-38df-0310-8fa8-9805f1628bb7