We always have to make one pass over the BOOTP options to see if the packet is BOOTP or DHCP; have "bootp_option()" take a flag indicating whether it's the first pass or the second and, in the first pass, don't put anything in the protocol tree - just return, through pointers, the DHCP packet type and the vendor class ID.  On the second pass, don't modify what those pointers point to, just use them as appropriate.
Make the vendor class ID pointer local to "dissect_bootp()", and have
"dissect_bootp()" do the first pass, set the Protocol and Info columns
appropriately if the packet is DHCP, and then, if a protocol tree is to
be built, make a second pass, handing "bootp_option()" the DHCP packet
type and vendor class ID we found.

If the vendor class ID is null, don't try to compare it with Intel's PXE
client indicator.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@3449 f5534014-38df-0310-8fa8-9805f1628bb7