Display all the digits of the microsecond field of an absolute time value.
Check that the microseconds field of an absolute time is valid, if it's
present.

Set "tm_isdst" in the "struct tm" handed to "mktime()" to -1, so that
"mktime()" will attempt to figure out whether the time is daylight
savings time or not.

Check that "mktime()" was able to convert the time.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@3487 f5534014-38df-0310-8fa8-9805f1628bb7