Don't base the protocol column setting in the WSP and WTP dissectors on the port - instead, base it on whether the dissector was called directly from UDP or called from another WAP dissector.
That way, if you explicitly say "decode this as WTP" because there was a
redirection (or if, in the future, the WSP dissector handles
redirections for you, although that won't handle the case of a capture
where the redirection wasn't captured), the column doesn't say "UDP", it
says the right thing.

Don't register the WTLS dissector by name - nobody calls it through a
handle.

Register the WTP dissector by name, as the WTLS dissector tries to get a
handle for it - although it doesn't actually call it, or the WSP
dissector, through a handle.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@3647 f5534014-38df-0310-8fa8-9805f1628bb7