"open_cap_file()" in Ethereal and Tethereal don't use the FILE_T they get from calling "wtap_file()", so get rid of the call and the (otherwise unused) variable to which its result gets assigned.
That lets us get rid of "wtap_file()" in Wiretap.

It also lets us get rid of the include of "zlib.h" in "file.h"; the
#defines of "file_open()", "filed_open()", and "file_close()" are also
unnecessary, so we get rid of those as well.

However, that means we need to include <zlib.h> in "gtk/main.c" and
"tethereal.c", so that the version number of libz is defined and can
show up in the version string.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@3652 f5534014-38df-0310-8fa8-9805f1628bb7