Put back the protocol tree fields for items with subtrees - but as FT_NONE, not FT_STRING, as they have no value assigned to them.
Don't use "hf_ssl_handshake_certificate" both for the collection of all
certificates and for individual certificates, add a new
"hf_ssl_handshake_certificates" FT_NONE field for the collection of all
certificates, as is done for other collections.

Properly pluralize the "N specs" in the item for the collection of
cipher specs.

Properly label the second "ssl.record" as "SSLv2 record data", not
"Length of SSLv2 record data".

Fix a typo in the field name of the field for the length of the
collection of certificates.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@3725 f5534014-38df-0310-8fa8-9805f1628bb7