Added a "Suppress Unmarked" option to the print dialog to allow you to suppress the printing of unmarked packets. This allows a user to mark the packets they wish to print and print ONLY those packets by suppressing all other unmarked packets.
This may seem like a bit of a convoluted way of expressing things,
as usually the desired behavior would be to print the marked packets.
However, we do NOT print marked packets that are not displayed under
the current filter.  To be maximally explicite I've expressed this
as suppressing unmarked frames.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@3736 f5534014-38df-0310-8fa8-9805f1628bb7