Loop over all the sub-payloads of a Security Association payload; make a common routine that loops over payloads.
Have that routine check for a payload of type "None", which means that
there's extra data after a payload that claimed to be the last one (by
virtue of having a next payload type of "None").

When dissecting a Security Association payload, make sure we have enough
data for the Domain of Interpretation field before putting it into the
tree, dissect the situation as a 4-byte quantity followed by a set of
sub-payloads only of the DOI is "IPSEC" (otherwise dissect it as raw
data), and make sure we have enough data for the 4-byte situation field
before putting it into the tree.

When dissecting a Proposal payload, show the raw bytes of the SPI.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@4081 f5534014-38df-0310-8fa8-9805f1628bb7