In "get_pointer_value()", don't include the terminating '\0' when formatting the string as text.
In "add_pointer_param()", make the offset in the text item for the
string the offset in the string area, not the offset of the pointer
itself.

Fix some items in item lists to have "lm_null" rather than a null
pointer, so that if we don't find an entry for the detail level, we
don't blow up dereferencing a null pointer, we just dissect all the
items based on the type in the descriptor string.

If we don't have a data descriptor for response data, just put in an
entry for the data; that may happen if we only dissected part of the
request to which the stuff we're dissecting is a reply, e.g. because the
snapshot length didn't let us capture all of the request.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@4085 f5534014-38df-0310-8fa8-9805f1628bb7