Update the code for the DHCP authentication option to match RFC 3118, and update the URL for the DHCP parameter value assignments; support a code of 90 for DHCP authentication, as that's what RFC 3118 specifies.
Show symbolic values for some of the fields in the DHCP authentication
option, and base the decision on whether to dissect the authentication
data as HMAC-MD5 data on whether that's what the protocol and algorithm
field specified.  Show the replay detection value as a 64-bit number
only if it's specified to be a monotonically-increasing counter;
otherwise, just show it as bytes.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@4116 f5534014-38df-0310-8fa8-9805f1628bb7