Attach a descriptive name field type and base to dissector tables; that specifies how the selector values used as keys in those tables are to be displayed, and the title to use when displaying the table.
Use that information in the code to display the initial and current
entries of various dissector tables.

Have the dissector for BACnet APDUs register itself by name, and have
the BACnet NPDU dissector call it iff the BAC_CONTROL_NET bit isn't set,
rather than doing it with a dissector table.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@4358 f5534014-38df-0310-8fa8-9805f1628bb7