Have routines to create GtkCTrees, set their line and expander style based on the user's UI preferences, and add them to a list of GtkCTrees. Use those routines to create all GtkCTrees.
Have a routine to update the preferences for all of those GtkCTrees.
Call that routine whenever the preferences change.

Label the line and expander style preferences as "Tree line style" and
"Tree expander style", as they no longer apply only to protocol trees.

Move the routines to maintain a list of scrolled windows, and to update
scrollbar placement for scrolled windows, to "ui_util.c".


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@4520 f5534014-38df-0310-8fa8-9805f1628bb7