Don't give tvbuffs names; instead, give data sources names, where a "data source" has a name and a top-level tvbuff, and frames can have a list of data sources associated with them.
Use the tvbuff pointer to determine which data source is the data source
for a given field; this means we don't have to worry about multiple data
sources with the same name - the only thing the name does is label the
notebook tab for the display of the data source, and label the hex dump
of the data source in print/Tethereal output.

Clean up a bunch of things discovered in the process of doing the above.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@4749 f5534014-38df-0310-8fa8-9805f1628bb7