Handle TPKT packets split across segment boundaries, and multiple TPKT packets per segment.
Instead of having a routine for dissectors such as the Q.931 dissector
to call to dissect the TPKT header, have a routine that does all the
reassembly and multiple-packets-per-segment work, and have the Q.931
dissector call it.  Export "is_tpkt()", and the new routine, to plugins.

Add preferences for TPKT and Q.931 reassembly.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@4778 f5534014-38df-0310-8fa8-9805f1628bb7