From Diana Eichert: add a "-q" flag to Tethereal to suppress packet count display.
Update the Tethereal man page to reflect the new option.

Update both the Ethereal and Tethereal man pages to use the same style
to describe options, e.g.

	-Z	Cause Ethereal to draw the mark of Zorro on the display.

rather than

	-Z	Causes Ethereal to draw the mark of Zorro on the display.

(some were using the first and some were using the second).

Update the Ethereal man page to do the same for menu items.

Update both the Ethereal and Tethereal man pages to better describe the
"-N" flag (by noting that any form of name resolution *not* specified in
the flag is turned *off*).


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@5005 f5534014-38df-0310-8fa8-9805f1628bb7