Start assigning ett_ values at 0, rather than 1; get rid of the reserved ETT_NONE entry.
Initialize the "tree_type" field of a "field_info" structure to -1,
meaning "this has not been given a subtree".  Add checks before using
that field that it's in range.  That way, you have to create a subtree
before putting protocol tree items under another item.

We allocate the "tree_is_expanded" array when we've registered all
dissectors; there's no need to allocate it while we're registering
dissectors and, in fact, doing so means we leak memory (the memory for
the version we allocated while registering dissectors).


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@5068 f5534014-38df-0310-8fa8-9805f1628bb7