Add an encapsulation type for "802.11 with radio information"; that type returns radio information such as signal strength, channel, and data rate in a pseudo-header.  Add that pseudo-header.
Use the "802.11 with radio information" encapsulation type for Wireless
Sniffer files; extract the radio information from where it appears to be
in the header.

Add dissector code for that encapsulation type.

Fix an error in the code to put radio information into the AiroPeek
tree.

Make the "wrapped" flag for NetXRay/Windows Sniffer captures a
"gboolean".


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@5122 f5534014-38df-0310-8fa8-9805f1628bb7