Use "CLEANUP_PUSH()" and "CLEANUP_CALL_AND_POP()" to plug potential memory leaks.
Put "Requests", not "X11 request", in the Info column for packets to the
server - we already know it's X11, and there may be more than one
request in the packet.

Put "Replies/events", not "X11 event", in the Info column for packets
from the server - we already know it's X11, and there may be more than
one message, and the messages might be replies rather than events.


git-svn-id: svn+ssh://svn.wireshark.org/svn/wireshark/trunk@5164 f5534014-38df-0310-8fa8-9805f1628bb7